Moved Schedule folder into SchedulingPage
