+++ autoscheduler/frontend/src/components/SchedulingPage/Schedule/AvailabilityCard/AvailabilityCard.css.d.ts
