+++ autoscheduler/frontend/src/components/SchedulingPage/Schedule/AvailabilityCard/AvailabilityCard.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { useDispatch } from 'react-redux';
 import DeleteIcon from '@material-ui/icons/Delete';
-import Availability from '../../../types/Availability';
+import Availability from '../../../../types/Availability';
 import ScheduleCard from '../ScheduleCard/ScheduleCard';
 import * as styles from './AvailabilityCard.css';
-import { deleteAvailability } from '../../../redux/actions/availability';
-import { setSelectedAvailability } from '../../../redux/actions/selectedAvailability';
-import { FIRST_HOUR, LAST_HOUR } from '../../../timeUtil';
+import { deleteAvailability } from '../../../../redux/actions/availability';
+import { setSelectedAvailability } from '../../../../redux/actions/selectedAvailability';
+import { FIRST_HOUR, LAST_HOUR } from '../../../../timeUtil';
 
 interface AvailabilityCardProps {
     availability: Availability;
