+++ autoscheduler/frontend/src/components/SchedulingPage/Schedule/HoveredTime/HoveredTime.css.d.ts
