+++ autoscheduler/frontend/src/components/SchedulingPage/Schedule/Schedule.css.d.ts
