+++ autoscheduler/frontend/src/components/SchedulingPage/Schedule/Schedule.tsx
@@ -1,16 +1,15 @@
 import * as React from 'react';
-import { RouteComponentProps } from '@reach/router';
 import { useSelector, useDispatch } from 'react-redux';
 import * as styles from './Schedule.css';
-import Meeting from '../../types/Meeting';
+import Meeting from '../../../types/Meeting';
 import MeetingCard from './MeetingCard/MeetingCard';
-import { RootState } from '../../redux/reducer';
-import { addAvailability, updateAvailability, mergeAvailability } from '../../redux/actions/availability';
-import { setSelectedAvailability, mergeThenSelectAvailability } from '../../redux/actions/selectedAvailability';
-import Availability, { AvailabilityType, AvailabilityArgs, roundUpAvailability } from '../../types/Availability';
+import { RootState } from '../../../redux/reducer';
+import { addAvailability, updateAvailability, mergeAvailability } from '../../../redux/actions/availability';
+import { setSelectedAvailability, mergeThenSelectAvailability } from '../../../redux/actions/selectedAvailability';
+import Availability, { AvailabilityType, AvailabilityArgs, roundUpAvailability } from '../../../types/Availability';
 import AvailabilityCard from './AvailabilityCard/AvailabilityCard';
 import HoveredTime from './HoveredTime/HoveredTime';
-import { FIRST_HOUR, LAST_HOUR, formatTime } from '../../timeUtil';
+import { FIRST_HOUR, LAST_HOUR, formatTime } from '../../../timeUtil';
 
 const Schedule: React.FC = () => {
   // these must be unique because of how they're used below
